Rename cutting.py to cutting_decomposition.py
