+++ circuit_knitting_toolbox/circuit_cutting/__init__.py
@@ -20,6 +20,6 @@ Circuit Cutting (:mod:`circuit_knitting_toolbox.circuit_cutting`).
 
     partition_circuit_qubits
 """
-from .cutting import partition_circuit_qubits
+from .cutting_decomposition import partition_circuit_qubits
 
 __all__ = ["partition_circuit_qubits"]
